Add files via upload
This is the initial file because the date range is from the start of the season to the day it is ran for the first time. The next version will update the database. In this case the start date will be a day after the last date stored in the db, retrieved by a query